refactor: Remove mutable_buffer crate dependency on query crate
